feat : add user info to navbar
