+++ src/App.jsx
@@ -4,9 +4,11 @@ import Dashboard from "./pages/Dashboard"
 import About from "./pages/About"
 import PSINavbar from "./components/PSINavbar"
 import { useEffect, useState } from "react"
+import axios from "axios"
 
 export default function App() {
   const [isLoggedIn, setIsLoggedIn] = useState(false)
+  const [userInfo, setUserInfo] = useState({})
 
   function handleLoggedIn(theval) {
     setIsLoggedIn(theval)
@@ -15,6 +17,19 @@ export default function App() {
   useEffect(() => {
     if (localStorage.getItem('token')) {
       handleLoggedIn(true)
+
+      const config = {
+        headers: {
+          Authorization: 'Bearer ' + localStorage.getItem('token')
+        }
+      }
+      axios.get(import.meta.env.VITE_APP_ENDPOINT + '/user', config)
+        .then((response) => {
+          const datanya = response.data
+          setUserInfo({ ...datanya })
+        }).catch(error => {
+          console.log({ at: 'app.jsx', error })
+        })
     } else {
       handleLoggedIn(false)
     }
@@ -23,11 +38,11 @@ export default function App() {
   return (
     <div>
       {
-        isLoggedIn ? <PSINavbar onLoggedIn={handleLoggedIn} /> : ''
+        isLoggedIn ? <PSINavbar onLoggedIn={handleLoggedIn} userInfo={userInfo} /> : ''
       }
       <Routes>
         <Route path="/" element={<Login onLoggedIn={handleLoggedIn} />} />
-        <Route path="/dashboard" element={<Dashboard />} />
+        <Route path="/dashboard" element={<Dashboard userInfo={userInfo} />} />
         <Route path="/about" element={<About />} />
         <Route path="*" element={<Login onLoggedIn={handleLoggedIn} />} />
       </Routes>
